+++ mavis/test/pages/sessions/sessions_search_page.py
@@ -2,7 +2,9 @@ from playwright.sync_api import Page, expect
 
 from mavis.test.annotations import step
 from mavis.test.models import (
+    Clinic,
     Programme,
+    School,
 )
 from mavis.test.pages.header_component import HeaderComponent
 from mavis.test.pages.sessions.search_component import SearchComponent
@@ -16,7 +18,7 @@ class SessionsSearchPage:
 
     @step("Click on {2} session at {1}")
     def click_session_for_programme_group(
-        self, location: str, programme_group: str
+        self, location: School | Clinic, programme_group: str
     ) -> None:
         if programme_group != Programme.MMR:
             for programme in Programme:
