+++ tests/test_children.py
@@ -1,8 +1,8 @@
 import pytest
 
-from mavis.test.accessibility import AccessibilityHelper
 from mavis.test.annotations import issue
 from mavis.test.data import ClassFileMapping, CohortsFileMapping, VaccsFileMapping
+from mavis.test.helpers.accessibility_helper import AccessibilityHelper
 from mavis.test.models import Programme
 from mavis.test.pages import (
     ChildActivityLogPage,
