+++ tests/test_consent_responses.py
@@ -1,9 +1,9 @@
 import pytest
 from playwright.sync_api import expect
 
-from mavis.test.accessibility import AccessibilityHelper
 from mavis.test.annotations import issue
 from mavis.test.data import ClassFileMapping, CohortsFileMapping, pds
+from mavis.test.helpers.accessibility_helper import AccessibilityHelper
 from mavis.test.models import Programme
 from mavis.test.pages import (
     ArchiveConsentResponsePage,
