+++ tests/test_import_child_lists.py
@@ -1,7 +1,7 @@
 import pytest
 
-from mavis.test.accessibility import AccessibilityHelper
 from mavis.test.data import ChildFileMapping
+from mavis.test.helpers.accessibility_helper import AccessibilityHelper
 from mavis.test.pages import (
     ChildrenSearchPage,
     DashboardPage,
