+++ tests/test_nurse_consent.py
@@ -1,8 +1,8 @@
 import pytest
 
-from mavis.test.accessibility import AccessibilityHelper
 from mavis.test.annotations import issue
 from mavis.test.data import CohortsFileMapping
+from mavis.test.helpers.accessibility_helper import AccessibilityHelper
 from mavis.test.models import ConsentMethod, DeliverySite, Programme, Vaccine
 from mavis.test.pages import (
     DashboardPage,
