+++ tests/test_online_consent_school_moves.py
@@ -1,7 +1,7 @@
 import pytest
 
-from mavis.test.accessibility import AccessibilityHelper
 from mavis.test.data import CohortsFileMapping
+from mavis.test.helpers.accessibility_helper import AccessibilityHelper
 from mavis.test.models import ConsentOption, Programme
 from mavis.test.pages import (
     ConsentResponsePage,
