+++ tests/test_school_moves.py
@@ -1,8 +1,8 @@
 import pytest
 from playwright.sync_api import expect
 
-from mavis.test.accessibility import AccessibilityHelper
 from mavis.test.data import ClassFileMapping
+from mavis.test.helpers.accessibility_helper import AccessibilityHelper
 from mavis.test.models import Programme
 from mavis.test.pages import (
     ChildActivityLogPage,
