+++ tests/test_start.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mavis.test.accessibility import AccessibilityHelper
+from mavis.test.helpers.accessibility_helper import AccessibilityHelper
 from mavis.test.pages import (
     StartPage,
 )
