+++ tests/test_vaccines.py
@@ -1,8 +1,8 @@
 import pytest
 from playwright.sync_api import expect
 
-from mavis.test.accessibility import AccessibilityHelper
 from mavis.test.annotations import issue
+from mavis.test.helpers.accessibility_helper import AccessibilityHelper
 from mavis.test.models import Vaccine
 from mavis.test.pages import (
     AddBatchPage,
